@@ -25,11 +25,22 @@
  *  });
  * ```
  */
-
+import '@mdi/font/css/materialdesignicons.css'
 import './index.css';
+import 'vuetify/styles';
 
 console.log('👋 This message is being logged by "renderer.js", included via Vite');
 import { createApp } from 'vue';
 import App from './App.vue';
+// Vuetify
 
-createApp(App).mount('#app');
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+
+const vuetify = createVuetify({ 
+  components,
+  directives,
+})
+createApp(App).use(vuetify).mount('#app');
