@@ -1,11 +1,35 @@
-import { defineConfig } from 'vite';
+import { defineConfig, mergeConfig } from 'vite';
+import {
+  getBuildConfig,
+  getBuildDefine,
+  external,
+  pluginHotRestart,
+} from './vite.base.config.mjs';
 
 // https://vitejs.dev/config
-export default defineConfig({
-  resolve: {
-    // Some libs that can run in both Web and Node.js, such as `axios`, we need to tell Vite to build them in Node.js.
-    browserField: false,
-    conditions: ['node'],
-    mainFields: ['module', 'jsnext:main', 'jsnext'],
-  },
+export default defineConfig((env) => {
+  /** @type {import('vite').ConfigEnv<'build'>} */
+  const forgeEnv = env;
+  const { forgeConfigSelf } = forgeEnv;
+  const define = getBuildDefine(forgeEnv);
+  const config = {
+    build: {
+      lib: {
+        entry: forgeConfigSelf.entry,
+        fileName: () => '[name].js',
+        formats: ['cjs'],
+      },
+      rollupOptions: {
+        external,
+      },
+    },
+    plugins: [pluginHotRestart('restart')],
+    define,
+    resolve: {
+      // Load the Node.js entry.
+      mainFields: ['module', 'jsnext:main', 'jsnext'],
+    },
+  };
+
+  return mergeConfig(getBuildConfig(forgeEnv), config);
 });
